Split lib/api.ts by domain
Old monolithic 617-line module becomes:
- lib/api/_internal.ts — endpoint resolution, fetchers, mappers, coalesced helper
- lib/api/polls.ts — poll CRUD + accessibility + related
- lib/api/multipolls.ts — multipoll CRUD + close/reopen/cutoff/thread-title
- lib/api/votes.ts — votes + POLL_VOTES_CHANGED_EVENT
- lib/api/results.ts — apiGetPollResults
- lib/api/search.ts — Nominatim/restaurant/movie/video-game search
- lib/api/index.ts — re-exports the public surface

Existing `from "@/lib/api"` import paths continue to work via the index.
